Move common evm util functions out of thorchain.ts
